Docstring and small changes
